+++ pokerapp/entity/entities.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+
+MessageId = str
+ChatId = str
+UserId = str
+Mention = str
+Score = int
+Money = int
