+++ pokerapp/entity/gamestate.py
@@ -1,0 +1,10 @@
+import enum
+
+
+class GameState(enum.Enum):
+    INITIAL = 0
+    ROUND_PRE_FLOP = 1  # No cards on the table.
+    ROUND_FLOP = 2  # Three cards.
+    ROUND_TURN = 3  # Four cards.
+    ROUND_RIVER = 4  # Five cards.
+    FINISHED = 5  # The end.
