+++ pokerapp/entity/playerstate.py
@@ -1,0 +1,7 @@
+import enum
+
+
+class PlayerState(enum.Enum):
+    ACTIVE = 1
+    FOLD = 0
+    ALL_IN = 10
