+++ pokerapp/entity/wallet.py
@@ -1,0 +1,34 @@
+from abc import abstractmethod
+
+from pokerapp.entity.entities import Money
+
+
+@abstractmethod
+class Wallet:
+    @staticmethod
+    def _prefix(id: int, suffix: str = ""):
+        pass
+
+    def add_daily(self) -> Money:
+        pass
+
+    def inc(self, amount: Money = 0) -> None:
+        pass
+
+    def inc_authorized_money(self, game_id: str, amount: Money) -> None:
+        pass
+
+    def authorized_money(self, game_id: str) -> Money:
+        pass
+
+    def authorize(self, game_id: str, amount: Money) -> None:
+        pass
+
+    def authorize_all(self, game_id: str) -> Money:
+        pass
+
+    def value(self) -> Money:
+        pass
+
+    def approve(self, game_id: str) -> None:
+        pass
